tests grob aufgesplittet
