+++ tests/testdata/querytests/company/companyABC.json
@@ -1,0 +1,1 @@
+{"name":"companyABC","address":"abc-address","city":"a-city","country":"de","boss":"bb-second-id"}
