+++ tests/testdata/querytests/company/companyZZZ.json
@@ -1,0 +1,1 @@
+{"name":"companyZZZ","address":"zzz-address","city":"a-city","country":"de","boss":"dd-fourth-id"}
