+++ tests/testdata/querytests/company/evil-company.json
@@ -1,0 +1,1 @@
+{"name":"evil-company","address":"666-address","city":"666-city","country":"hell","boss":"cc-third-id"}
