+++ tests/testdata/querytests/country/de.json
@@ -1,0 +1,1 @@
+{"short":"de","long":"germany","overlord":"bb-second-id","description":"germans live here"}
